user management

+++ application/controllers/settings.php
@@ -18,6 +18,7 @@ class Settings extends MY_Controller {
         $this->layout = 'main_layout.php';
         $this->load->model('dx_auth/users', 'users');
         $this->load->model('dx_auth/roles', 'roles');
+        $this->load->model('station_model');
         $this->load->model('dx_auth/user_profile', 'user_profile');
     }
 
@@ -34,7 +35,30 @@ class Settings extends MY_Controller {
      *  
      */
     public function users() {
-        $data['users'] = $this->users->get_users()->result();
+        $data['current_role'] = $currentRoleID = $this->session->userdata['DX_role_id'];
+        $data['is_ajax'] = false;
+        $roles = $this->roles->get_roles_list($currentRoleID)->result();
+        $params = null;
+        if (isAjax()) {
+            $data['is_ajax'] = true;
+            $params = array('station_id' => $this->input->post('station_id'), 'role_id' => $this->input->post('role_id'));
+        }
+
+        $data['users'] = $this->users->get_users($currentRoleID, $params)->result();
+        $data['roles'][''] = 'Select';
+        $data['stations'][''] = 'Select';
+        foreach ($roles as $value) {
+            $data['roles'][$value->id] = $value->name;
+        }
+        $stations = $this->station_model->get_all();
+
+        foreach ($stations as $value) {
+            $data['stations'][$value->id] = $value->station_name;
+        }
+        if (isAjax()) {
+            echo $this->load->view('settings/user', $data, TRUE);
+            exit;
+        }
         $this->load->view('settings/user', $data);
     }
 
@@ -62,48 +86,45 @@ class Settings extends MY_Controller {
     public function add_user() {
         $this->layout = 'default.php';
         $val = $this->form_validation;
+        $currentRoleID = $this->session->userdata['DX_role_id'];
+        if ($currentRoleID == 1 || $currentRoleID == 2 || $currentRoleID == 3) {
+            $val->set_rules('email', 'Email', 'trim|required|xss_clean|valid_email|callback_email_check');
+            $val->set_rules('password', 'Password', 'required');
+            $val->set_rules('first_name', 'First Name', 'trim|required|xss_clean');
+            $val->set_rules('last_name', 'Last Name', 'trim|required|xss_clean');
+            $val->set_rules('phone_no', 'Phone #', 'trim|xss_clean');
+            $val->set_rules('role', 'Role', 'trim|xss_clean|required');
 
-        $val->set_rules('email', 'Email', 'trim|required|xss_clean|valid_email|callback_email_check');
-        $val->set_rules('password', 'Password', 'required');
-        $val->set_rules('first_name', 'First Name', 'trim|required|xss_clean');
-        $val->set_rules('last_name', 'Last Name', 'trim|required|xss_clean');
-        $val->set_rules('phone_no', 'Phone #', 'trim|xss_clean');
-        $val->set_rules('role', 'Role', 'trim|xss_clean|required');
+            if ($this->input->post()) {
+                if ($val->run()) {
 
-        if ($this->input->post()) {
-            if ($val->run()) {
-
-                $record = array('email' => $val->set_value('email'),
-                    'password' => crypt($this->dx_auth->_encode($val->set_value('password'))),
-                    'role_id' => $val->set_value('role'),
-                );
-                $profile_data = array('first_name' => $val->set_value('first_name'),
-                    'last_name' => $val->set_value('last_name'),
-                    'phone_no' => $val->set_value('phone_no'),
-                );
-                $id = $this->users->create_user($record);
-                $profile_data['user_id'] = $id;
-                $this->user_profile->insert_profile($profile_data);
-                $this->session->set_userdata('saved', 'Record is Successfully Saved');
-                echo 'done';
-                exit;
-            } else {
-                $errors = $val->error_string();
-                $data['errors'] = $errors;
+                    $record = array('email' => $val->set_value('email'),
+                        'password' => crypt($this->dx_auth->_encode($val->set_value('password'))),
+                        'role_id' => $val->set_value('role'),
+                    );
+                    $profile_data = array('first_name' => $val->set_value('first_name'),
+                        'last_name' => $val->set_value('last_name'),
+                        'phone_no' => $val->set_value('phone_no'),
+                    );
+                    $id = $this->users->create_user($record);
+                    $profile_data['user_id'] = $id;
+                    $this->user_profile->insert_profile($profile_data);
+                    $this->session->set_userdata('saved', 'Record is Successfully Saved');
+                    echo 'done';
+                    exit;
+                } else {
+                    $errors = $val->error_string();
+                    $data['errors'] = $errors;
+                }
             }
+            $roles = $this->roles->get_roles_list($currentRoleID)->result();
+            foreach ($roles as $value) {
+                $data['roles'][$value->id] = $value->name;
+            }
+            echo $this->load->view('settings/add_user', $data, TRUE);
+        } else {
+            show_404();
         }
-
-
-
-
-        $superadmin = $this->session->userdata['DX_role_id'];
-
-
-        $roles = $this->roles->get_all($superadmin)->result();
-        foreach ($roles as $value) {
-            $data['roles'][$value->id] = $value->name;
-        }
-        echo $this->load->view('settings/add_user', $data, TRUE);
     }
 
     /**
@@ -114,51 +135,52 @@ class Settings extends MY_Controller {
     public function edit_user() {
         $this->layout = 'default.php';
         $user_id = $this->uri->segment(3);
-        $val = $this->form_validation;
+        $currentRoleID = $this->session->userdata['DX_role_id'];
+        if ($currentRoleID == 1 || $currentRoleID == 2 || $currentRoleID == 3) {
+            $val = $this->form_validation;
 
-        $val->set_rules('email', 'Email', 'trim|required|xss_clean|valid_email|callback_email_check[' . $user_id . ']');
-        $val->set_rules('password', 'Password', 'trim|xss_clean');
-        $val->set_rules('first_name', 'First Name', 'trim|required|xss_clean');
-        $val->set_rules('last_name', 'Last Name', 'trim|required|xss_clean');
-        $val->set_rules('phone_no', 'Phone #', 'trim|xss_clean');
-        $val->set_rules('role', 'Role', 'trim|xss_clean|required');
+            $val->set_rules('email', 'Email', 'trim|required|xss_clean|valid_email|callback_email_check[' . $user_id . ']');
+            $val->set_rules('password', 'Password', 'trim|xss_clean');
+            $val->set_rules('first_name', 'First Name', 'trim|required|xss_clean');
+            $val->set_rules('last_name', 'Last Name', 'trim|required|xss_clean');
+            $val->set_rules('phone_no', 'Phone #', 'trim|xss_clean');
+            $val->set_rules('role', 'Role', 'trim|xss_clean|required');
 
-        if ($this->input->post()) {
-            if ($val->run()) {
+            if ($this->input->post()) {
+                if ($val->run()) {
 
-                $record = array('email' => $val->set_value('email'),
-                    'role_id' => $val->set_value('role'),
-                );
-                if ($val->set_value('password') != '')
-                    $record['password'] = crypt($this->dx_auth->_encode($val->set_value('password')));
+                    $record = array('email' => $val->set_value('email'),
+                        'role_id' => $val->set_value('role'),
+                    );
+                    if ($val->set_value('password') != '')
+                        $record['password'] = crypt($this->dx_auth->_encode($val->set_value('password')));
 
-                $profile_data = array('first_name' => $val->set_value('first_name'),
-                    'last_name' => $val->set_value('last_name'),
-                    'phone_no' => $val->set_value('phone_no'),
-                );
-                $this->users->set_user($user_id, $record);
+                    $profile_data = array('first_name' => $val->set_value('first_name'),
+                        'last_name' => $val->set_value('last_name'),
+                        'phone_no' => $val->set_value('phone_no'),
+                    );
+                    $this->users->set_user($user_id, $record);
 
-                $this->user_profile->set_profile($user_id, $profile_data);
-                $this->session->set_userdata('updated', 'Record is Successfully Updated');
-                echo 'done';
-                exit;
-            } else {
-                $errors = $val->error_string();
-                $data['errors'] = $errors;
+                    $this->user_profile->set_profile($user_id, $profile_data);
+                    $this->session->set_userdata('updated', 'Record is Successfully Updated');
+                    echo 'done';
+                    exit;
+                } else {
+                    $errors = $val->error_string();
+                    $data['errors'] = $errors;
+                }
             }
+            $data['user_info'] = $this->users->get_user_detail($user_id)->row();
+
+
+            $roles = $this->roles->get_roles_list($currentRoleID)->result();
+            foreach ($roles as $value) {
+                $data['roles'][$value->id] = $value->name;
+            }
+            echo $this->load->view('settings/edit_user', $data, TRUE);
+        } else {
+            show_404();
         }
-        $data['user_info'] = $this->users->get_user_detail($user_id)->row();
-
-        $superadmin = $this->session->userdata['DX_role_id'];
-
-        $roles = $this->roles->get_all($superadmin)->result();
-        foreach ($roles as $value) {
-            $data['roles'][$value->id] = $value->name;
-        }
-
-
-
-        echo $this->load->view('settings/edit_user', $data, TRUE);
     }
 
     /**
@@ -168,9 +190,20 @@ class Settings extends MY_Controller {
      */
     public function delete_user() {
         $id = $this->uri->segment(3);
-        $delete_user = $this->users->delete_user($id);
-        $this->session->set_userdata('deleted', 'Record is Successfully Deleted');
-        redirect('settings/index', 'location');
+        $currentRoleID = $this->session->userdata['DX_role_id'];
+        if ($currentRoleID == 1 || $currentRoleID == 2 || $currentRoleID == 3) {
+            $currentUserID = $this->session->userdata['DX_user_id'];
+            if ($currentUserID != $id) {
+                $delete_user = $this->users->delete_user($id);
+                $this->session->set_userdata('deleted', 'Record is Successfully Deleted');
+            } else {
+                $this->session->set_userdata('deleted', 'You cannot delete currently active user.');
+            }
+
+            redirect('settings/index', 'location');
+        }
+        else
+            show_404();
     }
 
     /**
